Adding more routes

+++ controllers/feed.js
@@ -1,9 +1,9 @@
 const path = require('path');
 const fs = require('fs');
-
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -50,21 +50,31 @@ exports.createPost = (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
 
+    let creator;
+
     const post = new Post({
         title,
         imageUrl,
         content,
-        creator: {
-            name: 'Simon',
-        },
+        creator: req.userId,
     });
 
     post
         .save()
         .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.push(post);
+            creator = user;
+
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: "Create post successfully.",
-                post: result,
+                post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -127,6 +137,12 @@ exports.updatePost = (req, res, next) => {
                 throw error;
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorised');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if (post.imageUrl !== imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -159,11 +175,22 @@ exports.deletePost = (req, res, next) => {
                 throw error;
             }
 
-            // check user rights
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorised');
+                error.statusCode = 403;
+                throw error;
+            }
 
             clearImage(post.imageUrl);
 
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId)
+            return user.save();
         })
         .then(result => {
             res.status(200).json({ message: 'Post is deleted.' });
